# RentalManagement/openapi.yaml
{
  "openapi": "3.0.1",
  "info": {
    "title": "RentalManagement API",
    "description": "An ASP.NET Core Web API for managing rentals",
    "version": "v1"
  },
  "paths": {
...
    "/api/Place/{placeId}/Reservations": {
      "get": {
        "tags": [
          "Reservations"
        ],
        "summary": "Gets all Reservations for a specific Place",
        "parameters": [
          {
            "name": "placeId",
            "in": "path",
            "description": "ID of the place to get reservations from",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The Reservations were found.",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReservationDTO"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReservationDTO"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReservationDTO"
                  }
                }
              }
            }
          },
          "404": {
            "description": "The Place was not found."
          }
        }
      },
      "post": {
        "tags": [
          "Reservations"
        ],
        "summary": "Creates a new Reservation",
        "parameters": [
          {
            "name": "placeId",
            "in": "path",
            "description": "ID of the place to create reservation on",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "description": "Reservation information",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateReservationDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateReservationDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateReservationDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ReservationDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReservationDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReservationDTO"
                }
              }
            }
          },
          "201": {
            "description": "The Reservation was created."
          },
          "400": {
            "description": "The Reservation is invalid."
          },
          "404": {
            "description": "The Place was not found."
          }
        }
      }
    },
    "/api/Place/{placeId}/Reservations/{reservationId}": {
      "get": {
        "tags": [
          "Reservations"
        ],
        "summary": "Gets a Reservation by ID",
        "parameters": [
          {
            "name": "placeId",
            "in": "path",
            "description": "ID of the place to get reservation from",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "reservationId",
            "in": "path",
            "description": "ID of the reservation",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The Reservation was found.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ReservationDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReservationDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReservationDTO"
                }
              }
            }
          },
          "404": {
            "description": "The Place was not found."
          }
        }
      },
      "put": {
        "tags": [
          "Reservations"
        ],
        "summary": "Updates a Reservation by ID",
        "parameters": [
          {
            "name": "placeId",
            "in": "path",
            "description": "ID of the place that the reservation is a part of",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "reservationId",
            "in": "path",
            "description": "ID of the reservation to update",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "description": "Updated reservation information",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateReservationDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateReservationDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateReservationDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The Reservation was updated.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ReservationDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReservationDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReservationDTO"
                }
              }
            }
          },
          "400": {
            "description": "The updated Reservation is invalid."
          },
          "404": {
            "description": "The Reservation and/or Place was not found."
          }
        }
      },
      "delete": {
        "tags": [
          "Reservations"
        ],
        "summary": "Deletes a Reservation by its ID",
        "parameters": [
          {
            "name": "placeId",
            "in": "path",
            "description": "ID of the place that the reservation is a part of",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "reservationId",
            "in": "path",
            "description": "ID of the reservation",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/IActionResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IActionResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/IActionResult"
                }
              }
            }
          },
          "204": {
            "description": "The Reservation was deleted."
          },
          "400": {
            "description": "The Reservation does not belong to the specified place."
          },
          "404": {
            "description": "The Reservation and/or the place was not found."
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateReservationDTO": {
        "type": "object",
        "properties": {
          "startDate": {
            "type": "string",
            "format": "date-time",
            "example": "2021-09-01T00:00:00"
          },
          "endDate": {
            "type": "string",
            "format": "date-time",
            "example": "2021-09-01T00:00:00"
          },
          "price": {
            "type": "number",
            "format": "float",
            "example": 100
          }
        },
        "additionalProperties": false
      },
...
      "IActionResult": {
        "type": "object",
        "additionalProperties": false
      },
...
      "ReservationDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32",
            "example": 1
          },
          "placeId": {
            "type": "integer",
            "format": "int32",
            "example": 1
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2021-09-01T00:00:00"
          },
          "startDate": {
            "type": "string",
            "format": "date-time",
            "example": "2021-09-01T00:00:00"
          },
          "endDate": {
            "type": "string",
            "format": "date-time",
            "example": "2021-09-01T00:00:00"
          },
          "status": {
            "type": "string",
            "nullable": true,
            "example": "Reserved"
          },
          "price": {
            "type": "number",
            "format": "float",
            "example": 100
          }
        },
        "additionalProperties": false
      },
...
      "UpdateReservationDTO": {
        "type": "object",
        "properties": {
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2021-09-01T00:00:00"
          },
          "startDate": {
            "type": "string",
            "format": "date-time",
            "example": "2021-09-01T00:00:00"
          },
          "endDate": {
            "type": "string",
            "format": "date-time",
            "example": "2021-09-01T00:00:00"
          },
          "status": {
            "type": "string",
            "nullable": true,
            "example": "Reserved"
          },
          "price": {
            "type": "number",
            "format": "float",
            "example": 100
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}